Add files via upload
Adding separate pages for the mall staff panel so that refresh and back button function work. 
Added search feature for mall objects. 
Added list of objects that are sold out and need removed from the mall.
Removed Delete function.
